feat: add styled badge
