+++ src/themes/theme.ts
@@ -81,6 +81,7 @@ export default createTheme({
         highlight: '#FFEACC',
         sidebarContainer: 'rgba(32,32,33, 0.2)',
         grey: colors.grey,
+        predefinedBadgeColor: colors.green[100],
         text: {
             primary: colors.grey[900],
             secondary: colors.grey[800],
